Tweak how dehyphenation of stext works.
Previously the FZ_STEXT_DEHYPHENATE flag caused us to spot hyphens
at the end of lines, to remove them, and to join lines.

This left us in the slightly uncomfortable position of having
lines with a strange bbox, and missing chars (the hyphens).

Instead, we modify the meaning of this flag.

During stext extraction, we will always now spot a soft-hyphen
as the last element on a line as meaning "hyphenated". We set
a flag on the line to tell us that that line should be joined
to the next one.

If the flag is set, we spot other hyphen types in the same
position too and treat them as soft-hyphens.

This preserves lines as being physically distinct, and keeps
all the entries in the stext structure.

The stext buffer creation ("flattening") process now looks at the
line flags and suppresses the hyphen/newline at that stage.

So the net effect of the process is the same after flattening,
but the actual stext structure is different.